# openapi.yaml
openapi: 3.0.0
info:
  title: greensolar API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3006/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /cotizacion:
    get:
      tags:
        - Cotizacion
      summary: Listar todos los Cotizacions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Cotizacions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Cotizacion
      summary: Crear Cotizacion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Cotizacion creado
  /cotizacion/{id}:
    get:
      tags:
        - Cotizacion
      summary: Obtener Cotizacion por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cotizacion encontrado
        '404':
          description: Cotizacion no encontrado
    patch:
      tags:
        - Cotizacion
      summary: Actualizar Cotizacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Cotizacion actualizado
    delete:
      tags:
        - Cotizacion
      summary: Eliminar Cotizacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cotizacion eliminado
  /cotizacion/{id}/upload-image:
    post:
      tags:
        - Cotizacion
      summary: Subir imagen a Cotizacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /instalador:
    get:
      tags:
        - Instalador
      summary: Listar todos los Instaladors
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Instaladors
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Instalador
      summary: Crear Instalador
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Instalador creado
  /instalador/{id}:
    get:
      tags:
        - Instalador
      summary: Obtener Instalador por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instalador encontrado
        '404':
          description: Instalador no encontrado
    patch:
      tags:
        - Instalador
      summary: Actualizar Instalador
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Instalador actualizado
    delete:
      tags:
        - Instalador
      summary: Eliminar Instalador
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instalador eliminado
  /instalador/{id}/upload-image:
    post:
      tags:
        - Instalador
      summary: Subir imagen a Instalador
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...